reduce path by settin up jsconfig, remove unused style and add index for each component
