+++ package.json
@@ -3,9 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build",
+    "test": "react-app-rewired test",
     "eject": "react-scripts eject",
     "lint:fix": "eslint --fix .",
     "prepare": "husky install"
@@ -40,15 +40,18 @@
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "@eslint/js": "^9.5.0",
+    "customize-cra": "^1.0.0",
     "eslint": "^8.57.0",
     "eslint-config-react-app": "^7.0.1",
+    "eslint-import-resolver-alias": "^1.1.2",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-react": "^7.34.3",
     "eslint-plugin-react-hooks": "^4.6.2",
     "globals": "^15.6.0",
     "husky": "^8.0.3",
     "lint-staged": "^15.2.7",
-    "prop-types": "^15.8.1"
+    "prop-types": "^15.8.1",
+    "react-app-rewired": "^2.2.1"
   },
   "lint-staged": {
     "**/*.{js,jsx}": [
