+++ src/App.jsx
@@ -1,19 +1,19 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
-import KFCalendar from './components/Dashboard/Calendar/Calendar/KFCalendar';
+import { KFCalendar } from '@components/Dashboard/Calendar';
+import HomeDashboard from '@components/Dashboard/HomeDashboard';
+import LandingLayout from '@components/shared/LandingLayout';
 
-import HomeDashboard from './components/Dashboard/HomeDashboard/HomeDashboard';
-import LandingLayout from './components/shared/LandingLayout';
-import { EventProvider } from './context/EventContext';
-import Activate from './pages/Activate/Activate';
-import Dashboard from './pages/Dashboard/Dashboard';
+import { EventProvider } from '@context/EventContext';
 
-import FeaturesPage from './pages/FeaturesPage';
-import ForgetPassword from './pages/ForgetPassword';
-import Home from './pages/Home';
-import Register from './pages/Register';
-import ResetPassword from './pages/ResetPassword';
-import Signin from './pages/Signin';
+import Activate from '@pages/Activate/Activate';
+import Dashboard from '@pages/Dashboard/Dashboard';
+import FeaturesPage from '@pages/FeaturesPage';
+import ForgetPassword from '@pages/ForgetPassword';
+import Home from '@pages/Home';
+import Register from '@pages/Register';
+import ResetPassword from '@pages/ResetPassword';
+import Signin from '@pages/Signin';
 
 const App = () => (
   <BrowserRouter>
