+++ src/components/Dashboard/Calendar/index.js
@@ -2,3 +2,4 @@ export { default as DayEvent } from './DayEvent';
 export { default as WeekEvent } from './WeekEvent';
 export { default as MonthEvent } from './MonthEvent';
 export { default as KFToolbar } from './Toolbar';
+export { default as KFCalendar } from './Calendar';
