+++ src/components/Dashboard/DashboardHeader/index.js
@@ -1,1 +1,3 @@
-export { default } from './DashboardHeader';
+import DashboardHeader from './DashboardHeader';
+
+export default DashboardHeader;
