+++ src/components/Dashboard/HomeDashboard/MessageNotification/MessageNotification.module.css
@@ -1,0 +1,38 @@
+.rounded {
+  border: 1px solid #ccc;
+  border-radius: 20px !important;
+  padding: 0.8rem;
+}
+
+.messages {
+  display: flex;
+  flex-direction: row;
+  column-gap: 0.8rem;
+  padding: 0.5rem;
+}
+
+.circleShape {
+  background-color: #019da4;
+  border-radius: 50%;
+  width: 30px;
+  height: 30px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  color: white;
+  font-weight: bold;
+  margin-right: 10px;
+  font-family: Roboto;
+  font-size: 16px;
+  font-weight: 500;
+  line-height: 18.75px;
+  text-align: center;
+}
+
+.message {
+  font-family: Roboto;
+  font-size: 16px;
+  font-weight: 500;
+  line-height: 18.75px;
+  text-align: left;
+}
