+++ src/components/Dashboard/HomeDashboard/MessageNotification/index.js
@@ -1,0 +1,3 @@
+import MessageNotification from './MessageNotification';
+
+export default MessageNotification;
