+++ src/components/Dashboard/HomeDashboard/PendingEvents/DayEvent/DayEvent.module.css
@@ -1,0 +1,60 @@
+.event-box {
+  width: 555px;
+  margin-bottom: 10px;
+  border-radius: 5px;
+  padding: 5px;
+  border-left: 3px solid var(--box-bg-color);
+  height: 100%;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  position: relative;
+  top: 0;
+  left: 0;
+}
+.content-wrapper {
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  margin-left: 25px;
+  margin-top: 2px;
+}
+
+.title {
+  border-radius: 0px 7px 7px 0px;
+  font-family: Roboto;
+  color: #081821;
+  font-size: 15px;
+  font-weight: 500;
+  line-height: 18.75px;
+  text-align: left;
+  margin-left: 15px;
+  position: absolute;
+  top: -8px;
+  margin-top: 28px;
+}
+
+.timing {
+  font-size: 12px;
+  color: #011d26;
+  font-weight: 200;
+  line-height: 16.41px;
+  text-align: center;
+}
+
+.circle {
+  width: 27px;
+  height: 27px;
+  background-color: var(--circle-bg-color);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 15px;
+  color: white;
+  border-radius: 50%;
+  position: absolute;
+  top: 0;
+  left: 0;
+  margin-left: 5px;
+  margin-top: 15px;
+}
