+++ src/components/Dashboard/HomeDashboard/PendingEvents/index.js
@@ -1,0 +1,3 @@
+import PendingEvents from './PendingEvents';
+
+export default PendingEvents;
