+++ src/components/Dashboard/HomeDashboard/WeeksOverview/DayCard/index.js
@@ -1,0 +1,3 @@
+import DayCard from './DayCard';
+
+export default DayCard;
