+++ src/components/Dashboard/HomeDashboard/WeeksOverview/index.js
@@ -1,0 +1,3 @@
+import WeeksOverview from './WeeksOverview';
+
+export default WeeksOverview;
