+++ src/components/Dashboard/HomeDashboard/index.js
@@ -1,0 +1,3 @@
+import HomeDashboard from './HomeDashboard';
+
+export default HomeDashboard;
