+++ src/pages/Register/EmailVerify.jsx
@@ -1,9 +1,11 @@
 import PropTypes from 'prop-types';
 import React, { useState } from 'react';
 
-import { resendEmailVerification } from '../../api';
-import FeedbackBlock from '../../components/shared/Feedback/FeedbackBlock';
-import envelopeImg from '../../media/icons/email_sent.png';
+import { resendEmailVerification } from '@api';
+
+import FeedbackBlock from '@components/shared/Feedback/FeedbackBlock';
+
+import envelopeImg from '@media/icons/email_sent.png';
 
 import styles from './EmailVerify.module.css';
